Rename module and move cache to internal package
