+++ pkg/ac/checker/checker_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/grafana/rbac-client-poc/pkg/ac/models"
+	"github.com/grafana/authlib/pkg/ac/models"
 )
 
 func Test_WildcardsFromPrefixes(t *testing.T) {
